moved profile controller to admin folder
